% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat.R
\name{mcmc_rhat}
\alias{mcmc_rhat}
\alias{mcmc_rhat.default}
\alias{mcmc_rhat.brmsfit}
\title{Create Rhat plot}
\usage{
mcmc_rhat(object, ..., size = NULL)

\method{mcmc_rhat}{default}(object, ..., size = NULL)

\method{mcmc_rhat}{brmsfit}(object, pars = NULL, ..., size = NULL)
}
\arguments{
\item{object}{A model fitted with \code{[brms::brm()}].}

\item{...}{Arguments passed to individual methods.}

\item{pars}{An optional character vector of parameter names. 
For \code{nuts_params} these will be NUTS sampler parameter 
names rather than model parameters. If pars is omitted 
all parameters are included.}
}
\description{
\code{mcmc_rhat()} is a generic.  The default method simply calls \code{\link[bayesplot:mcmc_rhat]{bayesplot::mcmc_rhat()}}.
An additional method is provided for brmsfit objects so we can avoid first extracting
the R-hat values before creating the plot.
}
